// src/android/src/types/profile.types.ts
export interface PortfolioItem {
  id: string;
  profileId: string;
  title: string;
  description: string;
  imageUrl: string;
  projectUrl: string;
  githubUrl: string;
  kaggleUrl: string;
  technologies: string[];
  category: string;
  aiModels: string[];
  problemSolved: string;
  startDate: string;
  endDate: string;
  createdAt: string;
  updatedAt: string;
}
// ...
export interface PortfolioItemFormValues {
  title: string;
  description: string;
  projectUrl: string;
  githubUrl: string;
  kaggleUrl: string;
  technologies: string[];
  category: string;
  aiModels: string[];
  problemSolved: string;
  startDate: string;
  endDate: string;
}